feat: add ability to run bot in group threads
Also, rename `users` db collection to `chats`.

+++ bot/database/local_database.py
@@ -2,18 +2,18 @@ import json
 import os
 
 from database import database
-from models.user import User
+from models.chat import Chat
 
 
 class LocalDatabase(database.Database):
     def __init__(self, database_dir: str = "db"):
         self.database_dir = database_dir
 
-    def load_user(self, chat_id: int) -> User:
+    def load_chat(self, chat_id) -> Chat:
         path = os.path.join(self.database_dir, f"{chat_id}.json")
 
         if not os.path.exists(path):
-            return User()
+            return Chat()
 
         with open(path, "r") as f:
             json_str = f.read()
@@ -21,19 +21,19 @@ class LocalDatabase(database.Database):
             # reading json
             data = json.loads(json_str)
 
-            # creating new user object
-            user = User()
+            # creating new chat object
+            chat = Chat()
 
-            # adding dict data to the created user object
-            user.__dict__.update(data)
-            return user
+            # adding dict data to the created chat object
+            chat.__dict__.update(data)
+            return chat
 
-    def save_user(self, chat_id: int, user: User) -> None:
+    def save_chat(self, chat_id, chat) -> None:
         path = os.path.join(self.database_dir, f"{chat_id}.json")
 
         with open(path, "w") as f:
-            # getting dict from the user object
-            data = user.__dict__
+            # getting dict from the chat object
+            data = chat.__dict__
 
             # converting it into JSON
             json_str = json.dumps(data, indent=4, ensure_ascii=False)
@@ -41,8 +41,15 @@ class LocalDatabase(database.Database):
             # writing to the file
             f.write(json_str)
 
-    def clear_user(self, chat_id: int) -> None:
+    def clear_chat(self, chat_id, message_thread_id) -> None:
         path = os.path.join(self.database_dir, f"{chat_id}.json")
 
         if os.path.exists(path):
-            os.remove(path)
+            if message_thread_id is None:
+                os.remove(path)
+            else:
+                message_thread_id_str = str(message_thread_id)
+                chat = self.load_chat(chat_id)
+                if message_thread_id_str in chat.threads:
+                    del chat.threads[message_thread_id_str]
+                    self.save_chat(chat_id, chat)
